fix: env() string-to-int config coercion for all numeric config paths
Laravel's env() always returns string|null. Code that used is_int()
alone silently ignored user-configured values from environment variables
and fell back to defaults. This fix adds is_numeric() coercion fallbacks
to all numeric config paths:

- EventManager::getTriggerCacheTtl() — now accepts '0' string to disable
- DispatchTriggerJob::__construct() — retry.tries from env string
- GetsWebhookTimeout::getWebhookTimeout() — timeout from env string
- WebhookAction::getMaxFailures() — max_failures from env string
- Subscription::hasExceededFailures() — string config coercion
- Subscription::scopeExceededFailures() — string config coercion
- SubscriptionBuilder::save() — secret_length from env string
- EventScheduler::registerLogPurge() — simplified to (float) cast

Added EnvStringConfigCoercionTest (40+ tests) covering all coercion paths,
ConditionEngine between edge cases, sanitizePayloadForQueue nested structures,
DomainEvent::fromArray edge cases, WildcardMatcher edge cases, parseActions
edge cases, and null/not_null operators.

Bumped version to 5.77.0.

+++ src/EventScheduler.php
@@ -91,7 +91,8 @@ final class EventScheduler
     {
         $days = $this->getConfig()->get('events.retention.days');
 
-        if ($days === null || ! is_numeric($days) || (int) $days <= 0) {
+        // Accept int, numeric string, or float (e.g., 30.5 days)
+        if ($days === null || ! is_numeric($days) || (float) $days <= 0) {
             return;
         }
